requirements.txt
tqdm
pandas
pytorch
numpy
scikit-learn
pytorch-lightning
performer-pytorch==0.9
reformer-pytorch==1.2.3
linear-attention-transformer==0.15.0
routing_transformer==1.4.1
memformer==0.3.0
sinkhorn-transformer==0.11.1